% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{check_graple}
\alias{check_graple}
\title{Checks if the grapleObject is valid
by checking if the directories provided exist and valid}
\usage{
check_graple(object)
}
\arguments{
\item{object}{A Graple Object}
}
\value{
errors if any of the validation fails
}
\description{
Checks if the grapleObject is valid
by checking if the directories provided exist and valid
}
\examples{
\dontrun{
check_graple(grapleExp1)
}
}
\keyword{internal}

