% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{check_subdirectory}
\alias{check_subdirectory}
\title{Checks if the input/results directory is a sub-directory of
results/input directory}
\usage{
check_subdirectory(object)
}
\arguments{
\item{object}{A Graple Object}
}
\value{
A boolean value indicating whether the directory
structure is valid
}
\description{
Checks if the input/results directory is a sub-directory of
results/input directory
}
\examples{
\dontrun{
check_subdirectory(grapleExp1)
}
}
\keyword{internal}

