% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{createAPIKey}
\alias{createAPIKey}
\title{Used to create an API key that can be used as a parameter of a GRAPLE object.}
\usage{
createAPIKey(name, email)
}
\arguments{
\item{name}{A name for the user}

\item{email}{The user's email address}
}
\value{
The API key now associated with that user is returned, provided the email is not in use
}
\description{
Used to create an API key that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{
createAPIKey('Name','Email@test.com')
}
}
