% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{filePresent}
\alias{filePresent}
\title{checks if a given file exists in a directory}
\usage{
filePresent(dirPath, fileName)
}
\arguments{
\item{dirPath}{A Directory path}

\item{fileName}{File name}
}
\value{
A boolean value indicating whether a file is present or not
}
\description{
checks if a given file exists in a directory
}
\examples{
filePresent('C:/ExpRoot/Exp1/mySim', 'met_hourly.csv')
}
\keyword{internal}

