% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{filesPresent}
\alias{filesPresent}
\title{Checks if there are any files present in ExpRoot  Directory}
\usage{
filesPresent(object)
}
\arguments{
\item{object}{A Graple Object}
}
\value{
A boolean value indicating whether files are present or not
}
\description{
Checks if there are any files present in ExpRoot  Directory
}
\examples{
\dontrun{
filesPresent(grapleExp1)
}
}
\keyword{internal}

