% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{getExpID,Graple-method}
\alias{getExpID,Graple-method}
\title{Retrieves the experiment ID of the given GRAPLE object}
\usage{
\S4method{getExpID}{Graple}(grapleObject)
}
\arguments{
\item{grapleObject}{A GRAPLE object}
}
\value{
The experiment ID of the present GRAPLE object is returned
}
\description{
Retrieves the experiment ID of the given GRAPLE object
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
getExpID(grapleExp1)
}
}
