% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{getExpID}
\alias{getExpID}
\title{Retrieves the Experiment ID in order to prevent having to save the full Graple object.}
\usage{
getExpID(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
The Experiment ID is returned as a string and can be saved independently.
}
\description{
Retrieves the Experiment ID in order to prevent having to save the full Graple object.
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
getExpID(grapleObject)
}
}
