% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{getResultsDirName}
\alias{getResultsDirName}
\title{Gets the name for results directory}
\usage{
getResultsDirName(object)
}
\arguments{
\item{object}{A Graple Object}
}
\value{
ExpName if provided, else the JobID
}
\description{
Gets the name for results directory
}
\examples{
\dontrun{
getResultsDirName(grapleExp1)
}
}
\keyword{internal}

