% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{class}
\name{Graple-class}
\alias{Graple}
\alias{Graple-class}
\title{An S4 class to represent a graple object.}
\description{
An S4 class to represent a graple object.
}
\section{Slots}{

\describe{
\item{\code{GWSURL}}{A SubmissionURL for the experiment having a default value of http://graple.acis.ufl.edu}

\item{\code{ExpRootDir}}{Experiment Root Directory path}

\item{\code{ResultsDir}}{Directory path for storing the results}

\item{\code{JobID}}{Unique identifier for the experiment}

\item{\code{Email}}{Email address to send notifications}

\item{\code{APIKey}}{API Key to authenticate a user}

\item{\code{SimsPerJob}}{A number indicating the number of simulations bundled into a worker job}

\item{\code{StatusCode}}{Integer value indicating the status of an operation}

\item{\code{StatusMsg}}{A brief text message indicating the status of an operation}

\item{\code{ExpName}}{A name for the experiment}

\item{\code{TempDir}}{Temporary Directory path for temporary storage of files}

\item{\code{Retention}}{A user provided request to the GRAPLEr cluster on duration for which experiment results should be retained}

\item{\code{Client_Version_ID}}{The version of GRAPLEr package being used}
}}

