% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleAbortExperiment}
\alias{GrapleAbortExperiment}
\title{Aborts an existing GRAPLE experiment}
\usage{
GrapleAbortExperiment(submissionURL, experimentId)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}

\item{experimentId}{Experiment ID returned from GrapleRunExperiment
or GrapleRunExperimentSweep}
}
\value{
a status string
}
\description{
This function allows you to terminate a previously submitted experiment using
its experiment identifier. This is useful for aborting stalled or failed experiments.
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
expId<-"7YWMJYAYAR7Y3TNTAKC5801KMN7JHQW8NYBDMKUR"
GrapleAbortExperiment(graplerURL, expId)
}
}
\keyword{AbortExperiment}
\keyword{Graple}

