% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleCheckExperimentCompletion,Graple-method}
\alias{GrapleCheckExperimentCompletion,Graple-method}
\title{Checks the status of an experiment}
\usage{
\S4method{GrapleCheckExperimentCompletion}{Graple}(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Checks the status of an experiment
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleCheckExperimentCompletion(grapleExp1)
}
}

