% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleCheckService,Graple-method}
\alias{GrapleCheckService,Graple-method}
\title{Checks if the graple service is up and running on the submission URL provided}
\usage{
\S4method{GrapleCheckService}{Graple}(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Checks if the graple service is up and running on the submission URL provided
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleCheckService(grapleExp1)
}
}

