% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleCheckVersionCompatibility}
\alias{GrapleCheckVersionCompatibility}
\title{Checks version compatibility between R package and Graple web service}
\usage{
GrapleCheckVersionCompatibility(submissionURL)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}
}
\value{
true if versions are compatible else false
}
\description{
This function allows you to check version compatibility between R package and
Graple Web Service
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
GrapleCheckVersionCompatibility(graplerURL)
}
}
\keyword{CheckVersionCompatibility}
\keyword{Graple}

