% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleChkVersionCompatibility}
\alias{GrapleChkVersionCompatibility}
\title{Checks if the GRAPLEr pacakge being used is compatible with GWS Service being used}
\usage{
GrapleChkVersionCompatibility(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
A status msg indicating whether Web service and GRAPLEr are compatible
and the Graple object is returned
}
\description{
Checks if the GRAPLEr pacakge being used is compatible with GWS Service being used
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleChkVersionCompatibility(grapleObject)
}
}

