% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleCreateAPIKey,Graple-method}
\alias{GrapleCreateAPIKey,Graple-method}
\title{Used to create a class that can be used as a parameter of a GRAPLE object.}
\usage{
\S4method{GrapleCreateAPIKey}{Graple}(name, email)
}
\arguments{
\item{name}{A name for the user}

\item{email}{The user's email address}
}
\value{
The API key associated with that user is returned, provided the email is not in use
}
\description{
Used to create a class that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{

}
}
