% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleCreateClassKey,Graple-method}
\alias{GrapleCreateClassKey,Graple-method}
\title{Used to create a class that can be used as a parameter of a GRAPLE object.}
\usage{
\S4method{GrapleCreateClassKey}{Graple}(email, APIKey)
}
\arguments{
\item{email}{The user's email address}

\item{APIKey}{The user's APIkey}
}
\value{
The class key that is generated is returned
}
\description{
Used to create a class that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{
GrapleCreateClassKey('email@test.com', 'AP1K3YSAMPL3')
}
}
