% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleCreateClassKey}
\alias{GrapleCreateClassKey}
\title{Used to create a class that can be used as a parameter of a GRAPLE object.}
\usage{
GrapleCreateClassKey(email, APIKey)
}
\arguments{
\item{email}{The user's email address}

\item{APIKey}{The user's APIkey}
}
\value{
The class key that is generated is returned

The status message is updated on Graple object and the Graple object is returned
}
\description{
Used to create a class that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{
GrapleCreateClassKey('email@test.com', 'AP1K3YSAMPL3')
}
}
