% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleEndExperiment,Graple-method}
\alias{GrapleEndExperiment,Graple-method}
\title{A method to abort an experiment}
\usage{
\S4method{GrapleEndExperiment}{Graple}(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
A text message with the status of the operation
}
\description{
A method to abort an experiment
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleEndExperiment(grapleExp1)
}
}

