% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleEndExperiment}
\alias{GrapleEndExperiment}
\title{A method to abort an experiment}
\usage{
GrapleEndExperiment(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
A text message with the status of the operation
}
\description{
A method to abort an experiment
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleEndExperiment(grapleObject)
}
}

