% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleGetExperimentJobResults}
\alias{GrapleGetExperimentJobResults}
\title{Gets the GRAPLEr Experiment Job Results}
\usage{
GrapleGetExperimentJobResults(submissionURL, experimentId)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}

\item{experimentId}{Experiment ID returned from GrapleRunExperiment
or GrapleRunExperimentSweep}
}
\value{
a string describing the fully qualified result file name
}
\description{
This function allows you to retrieve the complete results
of an sweep job style experiment.
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
expId<-"7YWMJYAYAR7Y3TNTAKC5801KMN7JHQW8NYBDMKUR"
GrapleGetExperimentJobResults(graplerURL, expId)
}
}
\keyword{GetExperimentJobResults}
\keyword{Graple}

