% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleGetExperimentResults}
\alias{GrapleGetExperimentResults}
\title{Downloads the results to the results directory path with exp name/job id as the directory name}
\usage{
GrapleGetExperimentResults(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Downloads the results to the results directory path with exp name/job id as the directory name
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleGetExperimentResults(grapleObject)
}
}

