% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleGetPPOLibrary}
\alias{GrapleGetPPOLibrary}
\title{Retrieves the list of post process operation scripts}
\usage{
GrapleGetPPOLibrary(submissionURL)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}
}
\value{
a comma seperated
}
\description{
This function allows you to retrieve list of all the post-process operation scripts
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
GrapleGetPPOLibrary(graplerURL)
}
}
\keyword{GetPPOLibrary}
\keyword{Graple}

