% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleGetSimResult}
\alias{GrapleGetSimResult}
\title{Gets a Graple Simulation Results}
\usage{
GrapleGetSimResult(submissionURL, experimentId, simId)
}
\arguments{
\item{submissionURL}{URL:Port of the GrapeR service}

\item{experimentId}{Experiment ID returned from GrapleRunExperiment
or GrapleRunExperimentSweep}

\item{simId}{a valid simulation number from the sweep range}
}
\value{
a string describing the fully qualified result file name
}
\description{
This function allows you to retrieve the results of a
specific simulation in an experiment.
}
\examples{
\dontrun{
graplerURL<-"http://128.227.150.20:80"
expId<-"7YWMJYAYAR7Y3TNTAKC5801KMN7JHQW8NYBDMKUR"
simId<-2
GrapleGetSimResult(graplerURL, expId, simNum)
}
}
\keyword{GetSimResult}
\keyword{Graple}

