% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleListFilters}
\alias{GrapleListFilters}
\title{Retrieves the list of post process operation scripts}
\usage{
GrapleListFilters(submissionURL)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}
}
\value{
a comma seperated string of file names
}
\description{
This function allows you to retrieve list of all the post-process operation scripts
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
GrapleListFilters(graplerURL)
}
}
\keyword{Graple}
\keyword{ListFilters}

