% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleListPostProcessFilters}
\alias{GrapleListPostProcessFilters}
\title{Provides a list of post processing operations/ filters available}
\usage{
GrapleListPostProcessFilters(grapleObject)
}
\arguments{
\item{grapleObject}{A Graple Object}
}
\value{
Adds the list of filters to StatusMsg and returns the Graple object
}
\description{
Provides a list of post processing operations/ filters available
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
GrapleListPostProcessFilters(grapleObject)
}
}

