% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleRetrieveExperimentsByAPIKey,character-method}
\alias{GrapleRetrieveExperimentsByAPIKey,character-method}
\title{Used to retrieve the experiments associated with a given API key.}
\usage{
\S4method{GrapleRetrieveExperimentsByAPIKey}{character}(APIkey, url)
}
\arguments{
\item{APIkey}{An API key for the user}

\item{url}{The GWS url}
}
\value{
The experiments associated with that API key
}
\description{
Used to retrieve the experiments associated with a given API key.
}
\examples{
\dontrun{
GrapleRetrieveExperimentsByAPIKey('sampleAPIKey', 'GWS url')
}
}
