% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRetrieveExperimentsByAPIKey}
\alias{GrapleRetrieveExperimentsByAPIKey}
\title{Used to retrieve the experiments associated with a given API key.}
\usage{
GrapleRetrieveExperimentsByAPIKey(APIkey, url)
}
\arguments{
\item{url}{The GWS url}

\item{APIKey}{The user's APIkey}
}
\value{
The experiments associated with this API key
}
\description{
Used to retrieve the experiments associated with a given API key.
}
\examples{
\dontrun{
GrapleRetrieveExperimentsByAPIKey('AP1K3YSAMPL3', 'GWS url')
}
}
