% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleRetrieveExperimentsByClassKey,character-method}
\alias{GrapleRetrieveExperimentsByClassKey,character-method}
\title{Used to retrieve the experiments associated with a given class key.}
\usage{
\S4method{GrapleRetrieveExperimentsByClassKey}{character}(classKey, url)
}
\arguments{
\item{classKey}{The class key desired}

\item{url}{The GWS url}
}
\value{
The experiments associated with that class key
}
\description{
Used to retrieve the experiments associated with a given class key.
}
\examples{
\dontrun{
GrapleRetrieveExperimentsByClassKey('sampleClassKey', 'GWS url')
}
}
