% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRetrieveExperimentsByClassKey}
\alias{GrapleRetrieveExperimentsByClassKey}
\title{Used to retrieve the experiments associated with a given API key.}
\usage{
GrapleRetrieveExperimentsByClassKey(classKey, url)
}
\arguments{
\item{classKey}{The user's classKey}

\item{url}{The GWS url}
}
\value{
The experiments associated with this API key
}
\description{
Used to retrieve the experiments associated with a given API key.
}
\examples{
\dontrun{
GrapleRetrieveExperimentsByClassKey('AP1K3YSAMPL3', 'GWS url')
}
}
