% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRunExperimentSweep}
\alias{GrapleRunExperimentSweep}
\title{Creates an experiment based on sweep parameters}
\usage{
GrapleRunExperimentSweep(submissionURL, simDir, driverFileName, parameterName,
  startValue, endValue, numberOfIncrements, filterName)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}

\item{simDir}{the simulation folder containing the driver file}

\item{driverFileName}{the driver file name}

\item{parameterName}{the column name from the driver file to be swept}

\item{startValue}{numeric start offset}

\item{endValue}{numeric end offset}

\item{numberOfIncrements}{number of increments between start and end}

\item{filterName}{the name of the post-process filter (optional parameter)}
}
\value{
the experiment ID
}
\description{
This function allows you to run an experiment using a single simulation.
The sweep parameters are used to generate the other sims.
}
\examples{
\dontrun{
simDir="./Workspace/ExpRoot/Exp0"
driverFileName="met_hourly.csv"
parameterName="AirTemp"
startValue=-2
endValue=2
numberOfIncrements=10
filterName="Filter1.R" (optional parameter)
expId<-GrapleRunExperimentSweep(graplerURL, simDir, driverFileName, parameterName, startValue, endValue, numberOfIncrements)
}
}
\keyword{Graple}
\keyword{RunExperimentSweep}

