% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{setAPIKey,Graple-method}
\alias{setAPIKey,Graple-method}
\title{Sets the Class key in the grapleObject - can be done retroactively so that student can set their
class key}
\usage{
\S4method{setAPIKey}{Graple}(grapleObject, APIKey, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{APIKey}{A string API key (that would otherwise be in text file)}

\item{path}{Path to the a text file containing the API key}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the Class key in the grapleObject - can be done retroactively so that student can set their
class key
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setAPIKey(grapleExp1,'APIKey', 'path')
setAPIKey(grapleExp1,'APIKey')
}
}
