% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{setClassKey,Graple-method}
\alias{setClassKey,Graple-method}
\title{Sets the Class key in the grapleObject - can be done retroactively so that student can set their
class key/}
\usage{
\S4method{setClassKey}{Graple}(grapleObject, classKey, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{classKey}{A string class key (that would otherwise be in text file)}

\item{path}{Path to the a text file containing the class key}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the Class key in the grapleObject - can be done retroactively so that student can set their
class key/
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setClassKey(grapleExp1, 'classKeyText') # This would be if there was text of the class key
setClassKey(grapleExp1, 'C:/ExpRoot/KeyFiles/myClass.txt') # This would be if there was a file returned before
}
}
