% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setClassKey}
\alias{setClassKey}
\alias{setClassKey}
\title{Sets a class key in an existing GRAPLE object}
\usage{
setClassKey(grapleObject, classKey, path)

setClassKey(grapleObject, classKey, path)
}
\arguments{
\item{grapleObject}{The GRAPLE object for which you want to set the class key}

\item{classKey}{A textual class key}

\item{path}{A path to a class key file}

\item{grapleObject}{A Graple Object}

\item{path}{Path to the a text file containing the class key}

\item{classKey}{a textual class key}
}
\value{
The GRAPLE object with updated class key

The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets a class key in an existing GRAPLE object

Sets the class key in the grapleObject
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setAPIKey(grapleObject, 'C:/ExpRoot/KeyFiles/myKey.txt')
}
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setClassKey(grapleExp1, 'C:/ExpRoot/KeyFiles/myClass.txt') # This would be if there was a file returned before
setClassKey(grapleExp1, 'classKeyText') # This would be if there was text of the class key
}
}
