% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setExperimentDir}
\alias{setExperimentDir}
\title{Set the Experiment root directory in the Graple object}
\usage{
setExperimentDir(grapleObject, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{path}{Directory path for the experiment root directory}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Set the Experiment root directory in the Graple object
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setExperimentDir(grapleObject, 'C:/ExpRoot/Exp1')
}
}

