% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setExpName}
\alias{setExpName}
\title{Sets the name of the experiment in the Graple object}
\usage{
setExpName(grapleObject, expName)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{expName}{A Graple Object}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the name of the experiment in the Graple object
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setExpName(grapleObject, 'SweepExperiment')
}
}

