% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setResultsDir}
\alias{setResultsDir}
\title{Sets the results directory for the experiment in the Graple object}
\usage{
setResultsDir(grapleObject, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{path}{Directory path for the results to be downloaded}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the results directory for the experiment in the Graple object
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setResultsDir(grapleObject, 'C:/ExpRoot/Results/Exp1')
}
}

