% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setSecurityKey}
\alias{setSecurityKey}
\title{Sets the security key in the grapleObject}
\usage{
setSecurityKey(grapleObject, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{path}{Path to the a text file containing the security key}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the security key in the grapleObject
}
\examples{
\dontrun{
setSecurityKey(grapleExp1, 'C:/ExpRoot/KeyFiles/myKey.txt')
}
}

