% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setSubmissionURL}
\alias{setSubmissionURL}
\title{Sets the submission URL in the Graple object}
\usage{
setSubmissionURL(grapleObject, url)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{url}{A Submission URL}
}
\value{
The Graple object with updated submissionURL
}
\description{
Sets the submission URL in the Graple object
}
\examples{
\dontrun{
setSubmissionURL(grapleExp1, 'http://graple.acis.ufl.edu')
}
}

