% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{setTempDir}
\alias{setTempDir}
\title{Set the Temporary Directory to given directory path for the Graple Object}
\usage{
setTempDir(grapleObject, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{path}{Path to a directory to be set as Temporary directory for the experiment}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Set the Temporary Directory to given directory path for the Graple Object
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setTempDir(grapleObject, 'C:/TempDir')
}
}

