% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{validate_json}
\alias{validate_json}
\title{Checks if a json file is valid according to GRAPLEr constraints}
\usage{
validate_json(jsonFilePath)
}
\arguments{
\item{jsonFilePath}{JSON File path}
}
\value{
A boolean value indicating whether the json file is valid
}
\description{
Checks if a json file is valid according to GRAPLEr constraints
}
\examples{
validate_json('C:/ExpRoot/Exp3/job_desc.json')
}
\keyword{internal}

