% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{validate_url}
\alias{validate_url}
\title{Validates whether a given url exists}
\usage{
validate_url(url)
}
\arguments{
\item{url}{Submission URL}
}
\value{
A boolean value indicating whether url exists
}
\description{
Validates whether a given url exists
}
\examples{
validate_url('http://graple.acis.ufl.edu')
}
\keyword{internal}

