% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleCheckExperimentCompletion}
\alias{GrapleCheckExperimentCompletion}
\title{Check the Graple Experiment Status}
\usage{
GrapleCheckExperimentCompletion(submissionURL, experimentId)
}
\arguments{
\item{submissionURL}{URL:Port of the GrapleR service}

\item{experimentId}{Experiment ID returned from GrapleRunExperiment
or GrapleRunExperimentSweep}
}
\value{
a string describing experiment status
}
\description{
This function allows you to check the staus of the GrapleR service.
}
\examples{
\dontrun{
graplerURL<-"http://128.227.150.20:80"
expId<-"7YWMJYAYAR7Y3TNTAKC5801KMN7JHQW8NYBDMKUR"
GrapleCheckExperimentCompletion(graplerURL, expId)
}
}
\keyword{CheckExperimentCompletion}
\keyword{Graple}

