% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleCheckService}
\alias{GrapleCheckService}
\title{Get the Graple Service Status}
\usage{
GrapleCheckService(submissionURL)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}
}
\value{
a string describing service status
}
\description{
This function allows you to check the staus of the GRAPLEr web service.
}
\examples{
\dontrun{
GrapleCheckService()
}
}
\keyword{Graple}
\keyword{ServiceStatus}

