% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{GrapleCreateAPIKey,character-method}
\alias{GrapleCreateAPIKey,character-method}
\title{Used to create a class that can be used as a parameter of a GRAPLE object.}
\usage{
\S4method{GrapleCreateAPIKey}{character}(name, email, url)
}
\arguments{
\item{name}{A name for the user}

\item{email}{The user's email address}

\item{url}{The GWS url}
}
\value{
The API key associated with that user is returned, provided the email is not in use
}
\description{
Used to create a class that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{
GrapleCreateAPIKey('Name', 'Email@test.com', 'GWS url')
}
}
