% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleCreateAPIKey}
\alias{GrapleCreateAPIKey}
\title{Used to create an API key that can be used as a parameter of a GRAPLE object.}
\usage{
GrapleCreateAPIKey(name, email, url)
}
\arguments{
\item{name}{A name for the user}

\item{email}{The user's email address}

\item{url}{The GWS url}
}
\value{
The API key now associated with that user is returned, provided the email is not in use
}
\description{
Used to create an API key that can be used as a parameter of a GRAPLE object.
}
\examples{
\dontrun{
GrapleCreateAPIKey('Name','Email@test.com', 'GWS Url')
}
}
