% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleGetExperimentResults}
\alias{GrapleGetExperimentResults}
\title{Gets the Graple Experiment Results}
\usage{
GrapleGetExperimentResults(submissionURL, experimentId)
}
\arguments{
\item{submissionURL}{URL:Port of the GrapeR service}

\item{experimentId}{Experiment ID returned from GrapleRunExperiment
or GrapleRunExperimentSweep}
}
\value{
a string describing the fully qualified result file name
}
\description{
This function allows you to retrieve the complete results
of an experiment.
}
\examples{
\dontrun{
graplerURL<-"http://128.227.150.20:80"
expId<-"7YWMJYAYAR7Y3TNTAKC5801KMN7JHQW8NYBDMKUR"
GrapleGetExperimentResults(graplerURL, expId)
}
}
\keyword{GetExperimentResults}
\keyword{Graple}

