% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GraplePostProcessRunExperiment}
\alias{GraplePostProcessRunExperiment}
\title{Sends the experiment (multiple simulations) to be run on GrapleR
This function allows you to run graple.}
\usage{
GraplePostProcessRunExperiment(submissionURL, ExperimentDir, FilterName)
}
\arguments{
\item{submissionURL}{URL:Port of the GrapeR service}

\item{ExperimentDir}{the experiment root folder}

\item{FilterName}{the name of the post-processing filter script}
}
\examples{
\dontrun{
graplerURL<-"http://128.227.150.20:80"
expRootDir<-"./Workspace/SimRoot"
filterName<-"RunSimulation.R"
GraplePostProcessRunExperiment(graplerURL, expRootDir, filterName)
}
}
\keyword{Graple}
\keyword{RunExperiment}

