% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRunExperiment}
\alias{GrapleRunExperiment}
\title{Sends the experiment (multiple simulations) to be run on GrapleR
This function allows you to run graple.}
\usage{
GrapleRunExperiment(submissionURL, ExperimentDir)
}
\arguments{
\item{submissionURL}{URL:Port of the GrapeR service}

\item{ExperimentDir}{the experiment root folder}
}
\description{
Sends the experiment (multiple simulations) to be run on GrapleR
This function allows you to run graple.
}
\examples{
\dontrun{
graplerURL<-"http://128.227.150.20:80"
expRootDir<-"./Workspace/SimRoot"
GrapleRunExperiment(graplerURL, expRootDir)
}
}
\keyword{Graple}
\keyword{RunExperiment}

