% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRunExperiment}
\alias{GrapleRunExperiment}
\title{Sends the experiment (multiple simulations) to be run on GrapleR
This function allows you to run graple with an optional post-process filtering of results}
\usage{
GrapleRunExperiment(submissionURL, ExperimentDir, FilterName)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}

\item{ExperimentDir}{the experiment root folder}

\item{FilterName}{the name of the post-processing filter script}
}
\examples{
\dontrun{
graplerURL<-"http://graple-service.cloudapp.net"
expRootDir<-"./Workspace/ExpRoot"
filterName<-"Filter1.R"
GrapleRunExperiment(graplerURL, expRootDir, filterName)
}
}
\keyword{Graple}
\keyword{RunExperiment}

