% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\name{GrapleRunExperimentJob}
\alias{GrapleRunExperimentJob}
\title{Creates an experiment based on a configuration file}
\usage{
GrapleRunExperimentJob(submissionURL, simDir, FilterName)
}
\arguments{
\item{submissionURL}{URL:Port of the GRAPLEr web service}

\item{simDir}{the simulation folder containing the driver file}

\item{FilterName}{the name of post-process filter}
}
\value{
the experiment ID
}
\description{
This function allows you to run an experiment using a single simulation.
Create an archive that contains:
1. data csv file.
2. .nml file
3. job_desc.csv file (the name has to be job_desc.csv)
4. (optional) A FilterParams Directory, which consists of a "FilterParams.txt" file

The sweep parameters in the job description file are used to generate the other sims.
job_desc.csv is a comma separated value file in the following format:
base_file, driver_filename.csv
iterations, number_of_iterations
Parametername1, add, uniform, -1, 1
Paramtername2, sub, binomial, 10, 0.5
}
\examples{
\dontrun{
simDir="./Workspace/ExpRoot/Exp0"
FilterName="Filter1.R"
expId<-GrapleRunExperimentJob(graplerURL, simDir, FilterName)
}
}
\keyword{Graple}
\keyword{RunExperimentJob}

