% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapleR.R
\docType{methods}
\name{setAPIKey,Graple-method}
\alias{setAPIKey,Graple-method}
\title{Sets the API key in the grapleObject}
\usage{
\S4method{setAPIKey}{Graple}(grapleObject, path)
}
\arguments{
\item{grapleObject}{A Graple Object}

\item{path}{Path to the a text file containing the API key}
}
\value{
The status message is updated on Graple object and the Graple object is returned
}
\description{
Sets the API key in the grapleObject
}
\examples{
\dontrun{
grapleObject <- Graple(ExpRootDir="C:/InputDirectory", ResultsDir="C:/ResultsDirectory", TempDir = tempdir())
setAPIKey(grapleExp1, 'C:/ExpRoot/KeyFiles/myKey.txt')
}
}

